kusutkusut-backend/requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
selenium
Pillow
djangorestframework
django-cors-headers